+++ src/aks-preview/azext_aks_preview/custom.py
@@ -606,6 +606,7 @@ def aks_create(
     enable_cilium_dataplane=False,
     custom_ca_trust_certificates=None,
     enable_advanced_network_observability=None,
+    enable_fqdn_policy=None,
     # nodepool
     crg_id=None,
     message_of_the_day=None,
@@ -838,6 +839,8 @@ def aks_update(
     safeguards_excluded_ns=None,
     enable_advanced_network_observability=None,
     disable_advanced_network_observability=None,
+    enable_fqdn_policy=None,
+    disable_fqdn_policy=None,
     # metrics profile
     enable_cost_analysis=False,
     disable_cost_analysis=False,