+++ src/aks-preview/azext_aks_preview/tests/latest/test_managed_cluster_decorator.py
@@ -1009,6 +1009,71 @@ class AKSPreviewManagedClusterContextTestCase(unittest.TestCase):
         )
         self.assertEqual(ctx_5.get_enable_advanced_network_observability(), False)
 
+    def test_mc_get_enable_fqdn_policy(self):
+        # Default, not set.
+        ctx_1 = AKSPreviewManagedClusterContext(
+            self.cmd,
+            AKSManagedClusterParamDict({}),
+            self.models,
+            decorator_mode=DecoratorMode.CREATE,
+        )
+        self.assertEqual(ctx_1.get_enable_fqdn_policy(), None)
+
+        # Flag set to True.
+        ctx_2 = AKSPreviewManagedClusterContext(
+            self.cmd,
+            AKSManagedClusterParamDict(
+                {
+                    "enable_fqdn_policy": True,
+                }
+            ),
+            self.models,
+            decorator_mode=DecoratorMode.CREATE,
+        )
+        self.assertEqual(ctx_2.get_enable_fqdn_policy(), True)
+
+        # Flag set to True.
+        ctx_3 = AKSPreviewManagedClusterContext(
+            self.cmd,
+            AKSManagedClusterParamDict(
+                {
+                    "enable_fqdn_policy": True,
+                }
+            ),
+            self.models,
+            decorator_mode=DecoratorMode.UPDATE,
+        )
+        self.assertEqual(ctx_3.get_enable_fqdn_policy(), True)
+
+        # Flag set to True and False.
+        ctx_4 = AKSPreviewManagedClusterContext(
+            self.cmd,
+            AKSManagedClusterParamDict(
+                {
+                    "enable_fqdn_policy": True,
+                    "disable_fqdn_policy": True,
+                }
+            ),
+            self.models,
+            decorator_mode=DecoratorMode.UPDATE,
+        )
+        # fail on get_enable_network_observability mutual exclusive error
+        with self.assertRaises(MutuallyExclusiveArgumentError):
+            ctx_4.get_enable_advanced_network_observability()
+
+        # Flag set to False.
+        ctx_5 = AKSPreviewManagedClusterContext(
+            self.cmd,
+            AKSManagedClusterParamDict(
+                {
+                    "disable_fqdn_policy": True,
+                }
+            ),
+            self.models,
+            decorator_mode=DecoratorMode.UPDATE,
+        )
+        self.assertEqual(ctx_5.get_enable_fqdn_policy(), False)
+
     def test_get_enable_managed_identity(self):
         # custom value
         ctx_1 = AKSPreviewManagedClusterContext(
@@ -3914,7 +3979,7 @@ class AKSPreviewManagedClusterContextTestCase(unittest.TestCase):
             mode="Istio",
             istio=self.models.IstioServiceMesh(revisions=["asm-1-17", "asm-1-18"]),
         ))
-    
+
     def test_get_disable_local_accounts(self):
         # automatic cluster needs to enable the disable_local_accounts
         ctx_1 = AKSPreviewManagedClusterContext(
@@ -4429,7 +4494,7 @@ class AKSPreviewManagedClusterCreateDecoratorTestCase(unittest.TestCase):
         }
         automaticSKU = self.models.ManagedClusterSKU(name="Automatic", tier="Standard")
         ground_truth_mc_3 = self.models.ManagedCluster(
-            location="test_location", 
+            location="test_location",
             sku = automaticSKU,
             addon_profiles=addon_profiles_3,
             kind = "Automatic",
@@ -4659,7 +4724,7 @@ class AKSPreviewManagedClusterCreateDecoratorTestCase(unittest.TestCase):
         )
         automaticSKU = self.models.ManagedClusterSKU(name = "Automatic", tier = "Standard")
         ground_truth_mc_1 = self.models.ManagedCluster(
-            location = "test_location", 
+            location = "test_location",
             sku = automaticSKU,
             azure_monitor_profile = azure_monitor_profiles_1,
             kind = "Automatic",
@@ -8564,7 +8629,7 @@ class AKSPreviewManagedClusterUpdateDecoratorTestCase(unittest.TestCase):
         # fail on mutually exclusive uptime_sla and no_uptime_sla
         with self.assertRaises(MutuallyExclusiveArgumentError):
             dec_3.update_sku(mc_3)
-        
+
         dec_4 = AKSPreviewManagedClusterUpdateDecorator(
             self.cmd,
             self.client,
@@ -8809,7 +8874,7 @@ class AKSPreviewManagedClusterUpdateDecoratorTestCase(unittest.TestCase):
             self.cmd,
             self.client,
             {
-                "disable_static_egress_gateway": True, 
+                "disable_static_egress_gateway": True,
                 "enable_static_egress_gateway": True,
             },
             CUSTOM_MGMT_AKS_PREVIEW,
@@ -8921,7 +8986,7 @@ class AKSPreviewManagedClusterUpdateDecoratorTestCase(unittest.TestCase):
             self.cmd,
             self.client,
             {
-                "disable_imds_restriction": True, 
+                "disable_imds_restriction": True,
                 "enable_imds_restriction": True,
             },
             CUSTOM_MGMT_AKS_PREVIEW,